requirements.txt
pytest
Appium-Python-Client
selenium
webdriver_manager
allure-pytest
colorlog
